Update Engine.py
Minor fix, just added a comment to where return True should be in checkFood. This is so that this code will only check for food, but the actual code for what happens can easily be modified elsewhere. i.e. to play a sound, or something like that. Should be used by checking if it is true, and if it is true, then run code for eat in Snake class